feat: 🎸 add memeory configuration base on OS
improve the setup script to configured the memory and context lenght
based on the OS and the avialable memory on the sistem. As well this
commit solve half of the problem with the spinner, I am looking to
replace the UI with a lua package with some components but I want to
avoid add more dep to the project.